TEST DEPENDENCIES: Move to test asset directory, update tests & phpcs.xml
